+++ core/src/test/java/org/apache/iceberg/rest/TestErrorHandlers.java
@@ -20,7 +20,9 @@ package org.apache.iceberg.rest;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import org.apache.iceberg.exceptions.NoSuchWarehouseException;
 import org.apache.iceberg.exceptions.RESTException;
+import org.apache.iceberg.exceptions.ServiceFailureException;
 import org.apache.iceberg.rest.responses.ErrorResponse;
 import org.junit.jupiter.api.Test;
 
@@ -67,5 +69,39 @@ public class TestErrorHandlers {
     assertThatThrownBy(() -> ErrorHandlers.defaultErrorHandler().accept(error))
         .isInstanceOf(RESTException.class)
         .hasMessage("Unable to process (code: 422, type: ValidationException): null");
+  }
+
+  @Test
+  public void testConfigErrorHandler404ThrowsNoSuchWarehouseException() {
+    ErrorResponse error =
+        ErrorResponse.builder()
+            .responseCode(404)
+            .withType("NotFoundException")
+            .withMessage("Warehouse not found")
+            .build();
+
+    assertThatThrownBy(() -> ErrorHandlers.configErrorHandler().accept(error))
+        .isInstanceOf(NoSuchWarehouseException.class)
+        .hasMessage("Warehouse not found");
+  }
+
+  @Test
+  public void testConfigErrorHandler404ForMisconfiguredUri() {
+    ErrorResponse error =
+        ErrorResponse.builder().responseCode(404).withMessage("Not Found").build();
+
+    assertThatThrownBy(() -> ErrorHandlers.configErrorHandler().accept(error))
+        .isInstanceOf(RESTException.class)
+        .hasMessageContaining("Not Found");
+  }
+
+  @Test
+  public void testConfigErrorHandlerDelegatesToDefaultForNon404() {
+    ErrorResponse error =
+        ErrorResponse.builder().responseCode(500).withMessage("Internal server error").build();
+
+    assertThatThrownBy(() -> ErrorHandlers.configErrorHandler().accept(error))
+        .isInstanceOf(ServiceFailureException.class)
+        .hasMessageContaining("Internal server error");
   }
 }
