+++ core/src/test/java/org/apache/iceberg/rest/TestHTTPClient.java
@@ -50,9 +50,6 @@ import org.apache.hc.core5.http.HttpHost;
 import org.apache.hc.core5.http.HttpStatus;
 import org.apache.iceberg.IcebergBuild;
 import org.apache.iceberg.catalog.TableIdentifier;
-import org.apache.iceberg.exceptions.NoSuchWarehouseException;
-import org.apache.iceberg.exceptions.RESTException;
-import org.apache.iceberg.exceptions.ServiceFailureException;
 import org.apache.iceberg.relocated.com.google.common.collect.ImmutableMap;
 import org.apache.iceberg.rest.auth.AuthSession;
 import org.apache.iceberg.rest.auth.TLSConfigurer;
@@ -647,39 +644,5 @@ public class TestHTTPClient {
       Item item = (Item) o;
       return Objects.equals(id, item.id) && Objects.equals(data, item.data);
     }
-  }
-
-  @Test
-  public void testConfigErrorHandler404ThrowsNoSuchWarehouseException() {
-    ErrorResponse error =
-        ErrorResponse.builder()
-            .responseCode(404)
-            .withType("NotFoundException")
-            .withMessage("Warehouse not found")
-            .build();
-
-    assertThatThrownBy(() -> ErrorHandlers.configErrorHandler().accept(error))
-        .isInstanceOf(NoSuchWarehouseException.class)
-        .hasMessage("Warehouse not found");
-  }
-
-  @Test
-  public void testConfigErrorHandler404ForMisconfiguredUri() {
-    ErrorResponse error =
-        ErrorResponse.builder().responseCode(404).withMessage("Not Found").build();
-
-    assertThatThrownBy(() -> ErrorHandlers.configErrorHandler().accept(error))
-        .isInstanceOf(RESTException.class)
-        .hasMessageContaining("Not Found");
-  }
-
-  @Test
-  public void testConfigErrorHandlerDelegatesToDefaultForNon404() {
-    ErrorResponse error =
-        ErrorResponse.builder().responseCode(500).withMessage("Internal server error").build();
-
-    assertThatThrownBy(() -> ErrorHandlers.configErrorHandler().accept(error))
-        .isInstanceOf(ServiceFailureException.class)
-        .hasMessageContaining("Internal server error");
   }
 }
